refactor: rename const folder
